requirements.txt
neo4j>=5.0
requests>=2.0
python-dotenv>=1.0
pyyaml
chromadb>=0.5
sentence-transformers>=2.2

